PR #31602: Convert [DSA/RSA/Ecc]SecurityTransforms to new ASN.1 APIs
