+++ src/Common/src/System/Security/Cryptography/DSASecurityTransforms.cs
@@ -4,7 +4,9 @@
 
 using System.Diagnostics;
 using System.IO;
+using System.Numerics;
 using System.Security.Cryptography.Apple;
+using System.Security.Cryptography.Asn1;
 using Internal.Cryptography;
 
 namespace System.Security.Cryptography
@@ -77,6 +79,9 @@ namespace System.Security.Cryptography
 
                 public override DSAParameters ExportParameters(bool includePrivateParameters)
                 {
+                    // Apple requires all private keys to be exported encrypted, but since we're trying to export
+                    // as parsed structures we will need to decrypt it for the user.
+                    const string ExportPassword = "DotnetExportPassphrase";
                     SecKeyPair keys = GetKeys();
 
                     if (keys.PublicKey == null ||
@@ -85,33 +90,37 @@ namespace System.Security.Cryptography
                         throw new CryptographicException(SR.Cryptography_OpenInvalidHandle);
                     }
 
-                    DSAParameters parameters = new DSAParameters();
+                    byte[] keyBlob = Interop.AppleCrypto.SecKeyExport(
+                        includePrivateParameters ? keys.PrivateKey : keys.PublicKey,
+                        exportPrivate: includePrivateParameters,
+                        password: ExportPassword);
 
-                    DerSequenceReader publicKeyReader =
-                        Interop.AppleCrypto.SecKeyExport(keys.PublicKey, exportPrivate: false);
-
-                    publicKeyReader.ReadSubjectPublicKeyInfo(ref parameters);
-
-                    if (includePrivateParameters)
+                    try
                     {
-                        DerSequenceReader privateKeyReader =
-                            Interop.AppleCrypto.SecKeyExport(keys.PrivateKey, exportPrivate: true);
-
-                        privateKeyReader.ReadPkcs8Blob(ref parameters);
+                        if (!includePrivateParameters)
+                        {
+                            DSAKeyFormatHelper.ReadSubjectPublicKeyInfo(
+                                keyBlob,
+                                out int localRead,
+                                out DSAParameters key);
+                            Debug.Assert(localRead == keyBlob.Length);
+                            return key;
+                        }
+                        else
+                        {
+                            DSAKeyFormatHelper.ReadEncryptedPkcs8(
+                                keyBlob,
+                                ExportPassword,
+                                out int localRead,
+                                out DSAParameters key);
+                            Debug.Assert(localRead == keyBlob.Length);
+                            return key;
+                        }
                     }
-
-                    KeyBlobHelpers.TrimPaddingByte(ref parameters.P);
-                    KeyBlobHelpers.TrimPaddingByte(ref parameters.Q);
-
-                    KeyBlobHelpers.PadOrTrim(ref parameters.G, parameters.P.Length);
-                    KeyBlobHelpers.PadOrTrim(ref parameters.Y, parameters.P.Length);
-
-                    if (includePrivateParameters)
+                    finally
                     {
-                        KeyBlobHelpers.PadOrTrim(ref parameters.X, parameters.Q.Length);
+                        CryptographicOperations.ZeroMemory(keyBlob);
                     }
-
-                    return parameters;
                 }
 
                 public override void ImportParameters(DSAParameters parameters)
@@ -171,7 +180,41 @@ namespace System.Security.Cryptography
                 private static SafeSecKeyRefHandle ImportKey(DSAParameters parameters)
                 {
                     bool hasPrivateKey = parameters.X != null;
-                    byte[] blob = hasPrivateKey ? parameters.ToPrivateKeyBlob() : parameters.ToSubjectPublicKeyInfo();
+                    byte[] blob;
+                    
+                    if (hasPrivateKey)
+                    {
+                        Debug.Assert(parameters.X != null);
+
+                        // DSAPrivateKey ::= SEQUENCE(
+                        //   version INTEGER,
+                        //   p INTEGER,
+                        //   q INTEGER,
+                        //   g INTEGER,
+                        //   y INTEGER,
+                        //   x INTEGER,
+                        // )
+
+                        using (AsnWriter privateKeyWriter = new AsnWriter(AsnEncodingRules.DER))
+                        {
+                            privateKeyWriter.PushSequence();
+                            privateKeyWriter.WriteInteger(0);
+                            privateKeyWriter.WriteKeyParameterInteger(parameters.P);
+                            privateKeyWriter.WriteKeyParameterInteger(parameters.Q);
+                            privateKeyWriter.WriteKeyParameterInteger(parameters.G);
+                            privateKeyWriter.WriteKeyParameterInteger(parameters.Y);
+                            privateKeyWriter.WriteKeyParameterInteger(parameters.X);
+                            privateKeyWriter.PopSequence();
+                            blob = privateKeyWriter.Encode();
+                        }
+                    }
+                    else
+                    {
+                        using (AsnWriter writer = DSAKeyFormatHelper.WriteSubjectPublicKeyInfo(parameters))
+                        {
+                            blob = writer.Encode();
+                        }
+                    }
 
                     return Interop.AppleCrypto.ImportEphemeralKey(blob, hasPrivateKey);
                 }
@@ -286,155 +329,4 @@ namespace System.Security.Cryptography
 #if INTERNAL_ASYMMETRIC_IMPLEMENTATIONS
     }
 #endif
-
-    internal static class DsaKeyBlobHelpers
-    {
-        private static readonly Oid s_idDsa = new Oid("1.2.840.10040.4.1");
-
-        internal static void ReadSubjectPublicKeyInfo(this DerSequenceReader keyInfo, ref DSAParameters parameters)
-        {
-            // SubjectPublicKeyInfo::= SEQUENCE  {
-            //    algorithm AlgorithmIdentifier,
-            //    subjectPublicKey     BIT STRING  }
-            DerSequenceReader algorithm = keyInfo.ReadSequence();
-            string algorithmOid = algorithm.ReadOidAsString();
-
-            // EC Public Key
-            if (algorithmOid != s_idDsa.Value)
-            {
-                throw new CryptographicException();
-            }
-
-            // Dss-Parms ::= SEQUENCE {
-            //   p INTEGER,
-            //   q INTEGER,
-            //   g INTEGER
-            // }
-
-            DerSequenceReader algParameters = algorithm.ReadSequence();
-            byte[] publicKeyBlob = keyInfo.ReadBitString();
-            // We don't care about the rest of the blob here, but it's expected to not exist.
-
-            ReadSubjectPublicKeyInfo(algParameters, publicKeyBlob, ref parameters);
-        }
-
-        internal static void ReadSubjectPublicKeyInfo(
-            this DerSequenceReader algParameters,
-            byte[] publicKeyBlob,
-            ref DSAParameters parameters)
-        {
-            parameters.P = algParameters.ReadIntegerBytes();
-            parameters.Q = algParameters.ReadIntegerBytes();
-            parameters.G = algParameters.ReadIntegerBytes();
-
-            DerSequenceReader privateKeyReader = DerSequenceReader.CreateForPayload(publicKeyBlob);
-            parameters.Y = privateKeyReader.ReadIntegerBytes();
-
-            KeyBlobHelpers.TrimPaddingByte(ref parameters.P);
-            KeyBlobHelpers.TrimPaddingByte(ref parameters.Q);
-
-            KeyBlobHelpers.PadOrTrim(ref parameters.G, parameters.P.Length);
-            KeyBlobHelpers.PadOrTrim(ref parameters.Y, parameters.P.Length);
-        }
-
-        internal static byte[] ToSubjectPublicKeyInfo(this DSAParameters parameters)
-        {
-            // SubjectPublicKeyInfo::= SEQUENCE  {
-            //    algorithm AlgorithmIdentifier,
-            //    subjectPublicKey     BIT STRING  }
-
-            // Dss-Parms ::= SEQUENCE {
-            //   p INTEGER,
-            //   q INTEGER,
-            //   g INTEGER
-            // }
-
-            return DerEncoder.ConstructSequence(
-                DerEncoder.ConstructSegmentedSequence(
-                    DerEncoder.SegmentedEncodeOid(s_idDsa),
-                    DerEncoder.ConstructSegmentedSequence(
-                        DerEncoder.SegmentedEncodeUnsignedInteger(parameters.P),
-                        DerEncoder.SegmentedEncodeUnsignedInteger(parameters.Q),
-                        DerEncoder.SegmentedEncodeUnsignedInteger(parameters.G)
-                    )
-                ),
-                DerEncoder.SegmentedEncodeBitString(
-                    DerEncoder.SegmentedEncodeUnsignedInteger(parameters.Y))
-            );
-        }
-
-        internal static void ReadPkcs8Blob(this DerSequenceReader reader, ref DSAParameters parameters)
-        {
-            // Since the PKCS#8 blob for DSS/DSA does not include the public key (Y) this
-            // structure is only read after filling the public half.
-            Debug.Assert(parameters.P != null);
-            Debug.Assert(parameters.Q != null);
-            Debug.Assert(parameters.G != null);
-            Debug.Assert(parameters.Y != null);
-
-            // OneAsymmetricKey ::= SEQUENCE {
-            //   version                   Version,
-            //   privateKeyAlgorithm       PrivateKeyAlgorithmIdentifier,
-            //   privateKey                PrivateKey,
-            //   attributes            [0] Attributes OPTIONAL,
-            //   ...,
-            //   [[2: publicKey        [1] PublicKey OPTIONAL ]],
-            //   ...
-            // }
-            //
-            // PrivateKeyInfo ::= OneAsymmetricKey
-            //
-            // PrivateKey ::= OCTET STRING
-
-            int version = reader.ReadInteger();
-
-            // We understand both version 0 and 1 formats,
-            // which are now known as v1 and v2, respectively.
-            if (version > 1)
-            {
-                throw new CryptographicException();
-            }
-
-            {
-                // Ensure we're reading DSA, extract the parameters
-                DerSequenceReader algorithm = reader.ReadSequence();
-
-                string algorithmOid = algorithm.ReadOidAsString();
-
-                if (algorithmOid != s_idDsa.Value)
-                {
-                    throw new CryptographicException();
-                }
-
-                // The Dss-Params SEQUENCE is present here, but not needed since
-                // we got it from the public key already.
-            }
-
-            byte[] privateKeyBlob = reader.ReadOctetString();
-            DerSequenceReader privateKeyReader = DerSequenceReader.CreateForPayload(privateKeyBlob);
-            parameters.X = privateKeyReader.ReadIntegerBytes();
-        }
-
-        internal static byte[] ToPrivateKeyBlob(this DSAParameters parameters)
-        {
-            Debug.Assert(parameters.X != null);
-
-            // DSAPrivateKey ::= SEQUENCE(
-            //   version INTEGER,
-            //   p INTEGER,
-            //   q INTEGER,
-            //   g INTEGER,
-            //   y INTEGER,
-            //   x INTEGER,
-            // )
-
-            return DerEncoder.ConstructSequence(
-                DerEncoder.SegmentedEncodeUnsignedInteger(new byte[] { 0 }),
-                DerEncoder.SegmentedEncodeUnsignedInteger(parameters.P),
-                DerEncoder.SegmentedEncodeUnsignedInteger(parameters.Q),
-                DerEncoder.SegmentedEncodeUnsignedInteger(parameters.G),
-                DerEncoder.SegmentedEncodeUnsignedInteger(parameters.Y),
-                DerEncoder.SegmentedEncodeUnsignedInteger(parameters.X));
-        }
-    }
 }
