+++ src/Common/src/System/Security/Cryptography/KeyFormatHelper.cs
@@ -18,6 +18,7 @@ namespace System.Security.Cryptography
             string[] validOids,
             ReadOnlySpan<byte> source,
             KeyReader<TRet, TParsed> keyReader,
+            AsnEncodingRules keyEncodingRules,
             out int bytesRead,
             out TRet ret)
         {
@@ -25,7 +26,7 @@ namespace System.Security.Cryptography
             {
                 using (MemoryManager<byte> manager = new PointerMemoryManager<byte>(ptr, source.Length))
                 {
-                    ReadSubjectPublicKeyInfo(validOids, manager.Memory, keyReader, out bytesRead, out ret);
+                    ReadSubjectPublicKeyInfo(validOids, manager.Memory, keyReader, keyEncodingRules, out bytesRead, out ret);
                 }
             }
         }
@@ -52,6 +53,7 @@ namespace System.Security.Cryptography
             string[] validOids,
             ReadOnlyMemory<byte> source,
             KeyReader<TRet, TParsed> keyReader,
+            AsnEncodingRules keyEncodingRules,
             out int bytesRead,
             out TRet ret)
         {
@@ -76,7 +78,7 @@ namespace System.Security.Cryptography
                 // Fails if there are unconsumed bytes.
                 parsed = AsnSerializer.Deserialize<TParsed>(
                     spki.SubjectPublicKey,
-                    AsnEncodingRules.DER);
+                    keyEncodingRules);
             }
 
             keyReader(parsed, spki.Algorithm, out ret);
@@ -184,7 +186,7 @@ namespace System.Security.Cryptography
         {
             ReadEncryptedPkcs8(
                 validOids,
-                source.Span,
+                source,
                 password,
                 ReadOnlySpan<byte>.Empty,
                 keyReader,
@@ -208,31 +210,6 @@ namespace System.Security.Cryptography
                 keyReader,
                 out bytesRead,
                 out ret);
-        }
-
-        private static unsafe void ReadEncryptedPkcs8<TRet, TParsed>(
-            string[] validOids,
-            ReadOnlySpan<byte> source,
-            ReadOnlySpan<char> password,
-            ReadOnlySpan<byte> passwordBytes,
-            KeyReader<TRet, TParsed> keyReader,
-            out int bytesRead,
-            out TRet ret)
-        {
-            fixed (byte* ptr = &MemoryMarshal.GetReference(source))
-            {
-                using (MemoryManager<byte> manager = new PointerMemoryManager<byte>(ptr, source.Length))
-                {
-                    ReadEncryptedPkcs8(
-                        validOids,
-                        manager.Memory,
-                        password,
-                        passwordBytes,
-                        keyReader,
-                        out bytesRead,
-                        out ret);
-                }
-            }
         }
 
         private static void ReadEncryptedPkcs8<TRet, TParsed>(
