+++ src/System.Security.Cryptography.Algorithms/src/Internal/Cryptography/Helpers.cs
@@ -87,58 +87,5 @@ namespace Internal.Cryptography
             dest[2] = (byte)((value & 0xFF00) >> 8);
             dest[3] = (byte)(value & 0xFF);
         }
-
-        internal static byte[] ExportKeyParameter(this BigInteger value, int length)
-        {
-            byte[] target = new byte[length];
-
-            if (value.TryWriteBytes(target, out int bytesWritten, isUnsigned: true, isBigEndian: true))
-            {
-                if (bytesWritten < length)
-                {
-                    Buffer.BlockCopy(target, 0, target, length - bytesWritten, bytesWritten);
-                    target.AsSpan(0, length - bytesWritten).Clear();
-                }
-
-                return target;
-            }
-
-            throw new CryptographicException(SR.Cryptography_NotValidPublicOrPrivateKey);
-        }
-
-        internal static void WriteKeyParameterInteger(this AsnWriter writer, ReadOnlySpan<byte> integer)
-        {
-            Debug.Assert(!integer.IsEmpty);
-
-            if (integer[0] == 0)
-            {
-                int newStart = 1;
-
-                while (newStart < integer.Length)
-                {
-                    if (integer[newStart] >= 0x80)
-                    {
-                        newStart--;
-                        break;
-                    }
-
-                    if (integer[newStart] != 0)
-                    {
-                        break;
-                    }
-
-                    newStart++;
-                }
-
-                if (newStart == integer.Length)
-                {
-                    newStart--;
-                }
-
-                integer = integer.Slice(newStart);
-            }
-
-            writer.WriteIntegerUnsigned(integer);
-        }
     }
 }
