+++ src/System.Security.Cryptography.Algorithms/src/System/Security/Cryptography/DSA.cs
@@ -13,11 +13,6 @@ namespace System.Security.Cryptography
 {
     public abstract partial class DSA : AsymmetricAlgorithm
     {
-        private static readonly string[] s_validOids =
-        {
-            Oids.Dsa,
-        };
-
         public abstract DSAParameters ExportParameters(bool includePrivateParameters);
 
         public abstract void ImportParameters(DSAParameters parameters);
@@ -308,15 +303,7 @@ namespace System.Security.Cryptography
             {
                 try
                 {
-                    AsnWriter writer = new AsnWriter(AsnEncodingRules.DER);
-
-                    writer.PushSequence();
-                    writer.WriteInteger(0);
-                    WriteAlgorithmId(writer, dsaParameters);
-                    WriteKeyComponent(writer, dsaParameters.X, bitString: false);
-                    writer.PopSequence();
-
-                    return writer;
+                    return DSAKeyFormatHelper.WritePkcs8(dsaParameters);
                 }
                 finally
                 {
@@ -328,49 +315,7 @@ namespace System.Security.Cryptography
         private AsnWriter WriteSubjectPublicKeyInfo()
         {
             DSAParameters dsaParameters = ExportParameters(false);
-
-            AsnWriter writer = new AsnWriter(AsnEncodingRules.DER);
-
-            writer.PushSequence();
-            WriteAlgorithmId(writer, dsaParameters);
-            WriteKeyComponent(writer, dsaParameters.Y, bitString: true);
-            writer.PopSequence();
-
-            return writer;
-        }
-
-        private void WriteAlgorithmId(AsnWriter writer, in DSAParameters dsaParameters)
-        {
-            writer.PushSequence();
-            writer.WriteObjectIdentifier(Oids.Dsa);
-
-            // Dss-Parms ::= SEQUENCE {
-            //   p INTEGER,
-            //   q INTEGER,
-            //   g INTEGER  }
-            writer.PushSequence();
-            writer.WriteKeyParameterInteger(dsaParameters.P);
-            writer.WriteKeyParameterInteger(dsaParameters.Q);
-            writer.WriteKeyParameterInteger(dsaParameters.G);
-            writer.PopSequence();
-            writer.PopSequence();
-        }
-
-        private void WriteKeyComponent(AsnWriter writer, byte[] component, bool bitString)
-        {
-            using (AsnWriter inner = new AsnWriter(AsnEncodingRules.DER))
-            {
-                inner.WriteKeyParameterInteger(component);
-
-                if (bitString)
-                {
-                    writer.WriteBitString(inner.EncodeAsSpan());
-                }
-                else
-                {
-                    writer.WriteOctetString(inner.EncodeAsSpan());
-                }
-            }
+            return DSAKeyFormatHelper.WriteSubjectPublicKeyInfo(dsaParameters);
         }
 
         public override unsafe void ImportEncryptedPkcs8PrivateKey(
@@ -378,11 +323,9 @@ namespace System.Security.Cryptography
             ReadOnlySpan<byte> source,
             out int bytesRead)
         {
-            KeyFormatHelper.ReadEncryptedPkcs8<DSAParameters, DsaPrivateKeyAsn>(
-                s_validOids,
+            DSAKeyFormatHelper.ReadEncryptedPkcs8(
                 source,
                 passwordBytes,
-                ReadDsaPrivateKey,
                 out int localRead,
                 out DSAParameters ret);
 
@@ -406,11 +349,9 @@ namespace System.Security.Cryptography
             ReadOnlySpan<byte> source,
             out int bytesRead)
         {
-            KeyFormatHelper.ReadEncryptedPkcs8<DSAParameters, DsaPrivateKeyAsn>(
-                s_validOids,
+            DSAKeyFormatHelper.ReadEncryptedPkcs8(
                 source,
                 password,
-                ReadDsaPrivateKey, 
                 out int localRead,
                 out DSAParameters ret);
 
@@ -433,10 +374,8 @@ namespace System.Security.Cryptography
             ReadOnlySpan<byte> source,
             out int bytesRead)
         {
-            KeyFormatHelper.ReadPkcs8<DSAParameters, DsaPrivateKeyAsn>(
-                s_validOids,
+            DSAKeyFormatHelper.ReadPkcs8(
                 source,
-                ReadDsaPrivateKey, 
                 out int localRead,
                 out DSAParameters key);
 
@@ -455,67 +394,12 @@ namespace System.Security.Cryptography
             bytesRead = localRead;
         }
 
-        private void ReadDsaPrivateKey(
-            in DsaPrivateKeyAsn key,
-            in AlgorithmIdentifierAsn algId,
-            out DSAParameters ret)
-        {
-            if (!algId.Parameters.HasValue)
-            {
-                throw new CryptographicException(SR.Cryptography_Der_Invalid_Encoding);
-            }
-
-            DssParms parms =
-                AsnSerializer.Deserialize<DssParms>(algId.Parameters.Value, AsnEncodingRules.BER);
-
-            ret = new DSAParameters
-            {
-                P = parms.P.ToByteArray(isUnsigned: true, isBigEndian: true),
-                Q = parms.Q.ToByteArray(isUnsigned: true, isBigEndian: true),
-            };
-
-            ret.G = parms.G.ExportKeyParameter(ret.P.Length);
-
-            // Force a positive interpretation because Windows sometimes writes negative numbers.
-            BigInteger x = new BigInteger(key.X.Value.Span, isUnsigned: true, isBigEndian: true);
-            ret.X = x.ExportKeyParameter(ret.Q.Length);
-
-            // The public key is not contained within the format, calculate it.
-            BigInteger y = BigInteger.ModPow(parms.G, x, parms.P);
-            ret.Y = y.ExportKeyParameter(ret.P.Length);
-        }
-
-        private void ReadDsaPublicKey(
-            in BigInteger y,
-            in AlgorithmIdentifierAsn algId,
-            out DSAParameters ret)
-        {
-            if (!algId.Parameters.HasValue)
-            {
-                throw new CryptographicException(SR.Cryptography_Der_Invalid_Encoding);
-            }
-
-            DssParms parms =
-                AsnSerializer.Deserialize<DssParms>(algId.Parameters.Value, AsnEncodingRules.BER);
-
-            ret = new DSAParameters
-            {
-                P = parms.P.ToByteArray(isUnsigned: true, isBigEndian: true),
-                Q = parms.Q.ToByteArray(isUnsigned: true, isBigEndian: true),
-            };
-
-            ret.G = parms.G.ExportKeyParameter(ret.P.Length);
-            ret.Y = y.ExportKeyParameter(ret.P.Length);
-        }
-
         public override void ImportSubjectPublicKeyInfo(
             ReadOnlySpan<byte> source,
             out int bytesRead)
         {
-            KeyFormatHelper.ReadSubjectPublicKeyInfo<DSAParameters, BigInteger>(
-                s_validOids,
+            DSAKeyFormatHelper.ReadSubjectPublicKeyInfo(
                 source,
-                ReadDsaPublicKey,
                 out int localRead,
                 out DSAParameters key);
