+++ src/System.Security.Cryptography.Algorithms/src/System/Security/Cryptography/ECDsa.cs
@@ -263,7 +263,7 @@ namespace System.Security.Cryptography
             {
                 try
                 {
-                    using (AsnWriter pkcs8PrivateKey = ecParameters.WritePkcs8PrivateKey())
+                    using (AsnWriter pkcs8PrivateKey = EccKeyFormatHelper.WritePkcs8PrivateKey(ecParameters))
                     using (AsnWriter writer = KeyFormatHelper.WriteEncryptedPkcs8(
                         passwordBytes,
                         pkcs8PrivateKey,
@@ -299,7 +299,7 @@ namespace System.Security.Cryptography
             {
                 try
                 {
-                    using (AsnWriter pkcs8PrivateKey = ecParameters.WritePkcs8PrivateKey())
+                    using (AsnWriter pkcs8PrivateKey = EccKeyFormatHelper.WritePkcs8PrivateKey(ecParameters))
                     using (AsnWriter writer = KeyFormatHelper.WriteEncryptedPkcs8(
                         password,
                         pkcs8PrivateKey,
@@ -325,7 +325,7 @@ namespace System.Security.Cryptography
             {
                 try
                 {
-                    using (AsnWriter writer = ecParameters.WritePkcs8PrivateKey())
+                    using (AsnWriter writer = EccKeyFormatHelper.WritePkcs8PrivateKey(ecParameters))
                     {
                         return writer.TryEncode(destination, out bytesWritten);
                     }
@@ -343,7 +343,7 @@ namespace System.Security.Cryptography
         {
             ECParameters ecParameters = ExportParameters(false);
 
-            using (AsnWriter writer = ecParameters.WriteSubjectPublicKeyInfo())
+            using (AsnWriter writer = EccKeyFormatHelper.WriteSubjectPublicKeyInfo(ecParameters))
             {
                 return writer.TryEncode(destination, out bytesWritten);
             }
@@ -358,7 +358,7 @@ namespace System.Security.Cryptography
                 s_validOids,
                 source,
                 passwordBytes,
-                ECParameters.FromECPrivateKey,
+                EccKeyFormatHelper.FromECPrivateKey,
                 out int localRead,
                 out ECParameters ret);
 
@@ -385,7 +385,7 @@ namespace System.Security.Cryptography
                 s_validOids,
                 source,
                 password,
-                ECParameters.FromECPrivateKey,
+                EccKeyFormatHelper.FromECPrivateKey,
                 out int localRead,
                 out ECParameters ret);
 
@@ -410,7 +410,7 @@ namespace System.Security.Cryptography
             KeyFormatHelper.ReadPkcs8<ECParameters, ECPrivateKey>(
                 s_validOids,
                 source,
-                ECParameters.FromECPrivateKey,
+                EccKeyFormatHelper.FromECPrivateKey,
                 out int localRead,
                 out ECParameters key);
 
@@ -435,7 +435,8 @@ namespace System.Security.Cryptography
             KeyFormatHelper.ReadSubjectPublicKeyInfo<ECParameters, ReadOnlyMemory<byte>>(
                 s_validOids,
                 source,
-                ECParameters.FromECPublicKey,
+                EccKeyFormatHelper.FromECPublicKey,
+                AsnEncodingRules.DER,
                 out int localRead,
                 out ECParameters key);
 
@@ -445,7 +446,7 @@ namespace System.Security.Cryptography
 
         public virtual unsafe void ImportECPrivateKey(ReadOnlySpan<byte> source, out int bytesRead)
         {
-            ECParameters ecParameters = ECParameters.FromECPrivateKey(source, out int localRead);
+            ECParameters ecParameters = EccKeyFormatHelper.FromECPrivateKey(source, out int localRead);
 
             fixed (byte* privPin = ecParameters.D)
             {
@@ -469,7 +470,7 @@ namespace System.Security.Cryptography
             {
                 try
                 {
-                    using (AsnWriter writer = ecParameters.WriteECPrivateKey())
+                    using (AsnWriter writer = EccKeyFormatHelper.WriteECPrivateKey(ecParameters))
                     {
                         return writer.Encode();
                     }
@@ -489,7 +490,7 @@ namespace System.Security.Cryptography
             {
                 try
                 {
-                    using (AsnWriter writer = ecParameters.WriteECPrivateKey())
+                    using (AsnWriter writer = EccKeyFormatHelper.WriteECPrivateKey(ecParameters))
                     {
                         return writer.TryEncode(destination, out bytesWritten);
                     }
