+++ src/System.Security.Cryptography.X509Certificates/src/Internal/Cryptography/Pal.OSX/X509Pal.cs
@@ -4,8 +4,10 @@
 
 using System;
 using System.Diagnostics;
+using System.Numerics;
 using System.Security.Cryptography;
 using System.Security.Cryptography.Apple;
+using System.Security.Cryptography.Asn1;
 using System.Security.Cryptography.X509Certificates;
 
 namespace Internal.Cryptography.Pal
@@ -82,9 +84,13 @@ namespace Internal.Cryptography.Pal
 
             private static AsymmetricAlgorithm DecodeRsaPublicKey(byte[] encodedKeyValue)
             {
-                DerSequenceReader reader = new DerSequenceReader(encodedKeyValue);
-                RSAParameters rsaParameters = new RSAParameters();
-                reader.ReadPkcs1PublicBlob(ref rsaParameters);
+                RSAPublicKey publicKey = AsnSerializer.Deserialize<RSAPublicKey>(encodedKeyValue, AsnEncodingRules.BER);
+
+                AlgorithmIdentifierAsn ignored = default;
+                RSAKeyFormatHelper.ReadRsaPublicKey(
+                    publicKey,
+                    ignored,
+                    out RSAParameters rsaParameters);
 
                 RSA rsa = RSA.Create();
                 try
@@ -101,10 +107,12 @@ namespace Internal.Cryptography.Pal
 
             private static AsymmetricAlgorithm DecodeDsaPublicKey(byte[] encodedKeyValue, byte[] encodedParameters)
             {
-                DSAParameters dsaParameters = new DSAParameters();
-                DerSequenceReader parameterReader = new DerSequenceReader(encodedParameters);
+                BigInteger y = AsnSerializer.Deserialize<BigInteger>(encodedKeyValue, AsnEncodingRules.DER);
 
-                parameterReader.ReadSubjectPublicKeyInfo(encodedKeyValue, ref dsaParameters);
+                DSAKeyFormatHelper.ReadDsaPublicKey(
+                    y,
+                    new AlgorithmIdentifierAsn { Parameters = encodedParameters },
+                    out DSAParameters dsaParameters);
 
                 DSA dsa = DSA.Create();
                 try
